--- src/pages/ddd.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-const ddd = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default ddd
-
